refactor(docs-tests): extract pure route-check helpers for unit coverage
